requirements.txt
mysql-connector-python==8.0.27
protobuf==3.19.1
python-dotenv==0.19.2